refactor: extract auth into app/auth/ (Django-style app)
Third pilot, validating the pattern for a feature with no model file of
its own (User/tokens stay in app/models/ since courses.py/main.py
haven't been decomposed yet):

- app/blueprints/auth.py -> app/auth/routes.py
- app/services/auth_service.py -> app/auth/services.py (email_service.py
  stays put — it's used by superadmin/tickets/health too, genuinely
  shared rather than auth-owned)
- app/templates/auth/ -> app/auth/templates/auth/
- tests/test_auth.py -> app/auth/tests/ (no local conftest needed, only
  uses fixtures already in app/conftest.py)

Verified: create_app() builds with all 224 routes intact, pytest
unchanged from baseline (25 pre-existing failures, 99 passing).
