+++ app/auth/templates/auth/forgot_password.html
